[wip] re-implement hard stopping of CPUs
Hard stopping changes:
- stop_cpus_hard interface is intentionally narrowed to always act on all
  other cpus
- stop_cpus_hard keeps its own accounting of stopped cpus completely
  disregarding soft stopping of cpus (pausing, suspending)
- no recursion whatsoever is allowed for stop_cpus_hard; all callers
  must check if they are already in "super" mode
- protect stop function with spinlock context
- avoid potential deadlock between two cpus already being in nmi context
  (e.g. because of hardware) and trying to stop each other by checking
  the stopped bit while spinning on stopper lock; this is possibly an
  arch-specific thing [x86];

The last item is what I call a cross-cpu deadlock avoidance approach.
General idea: if a CPU can't grab a lock protecting cross-/intra-CPU
interactions, then while spinning on the lock the CPU should check
for incoming cross-CPU events possibly posted by a CPU that currently
holds the lock.  In this scheme an IPI is used only to interrupt a CPU
so that it can notice a new event, but the actual event should be passed
via the memory (a bit in a cpu set - like newly introduced
hard_stopping_cpus, or some other variable).

Only x86 is covered.

kgdb is updated to pay attention to hard_stopped_cpus as well...