uma: keep stats on zone working set sizes and periodically trim bucket caches
The stats are taken periodically from uma_timeout().  The stats consist
of previous and previous previous values of the bucket cache sizes.
When there are more than 12.5% of free cached items in a zone and its
backing keg the code tries to trim the bucket cache by freeing the minimum
of the previous sizes and the current size.  It is assumed that that many
items are outside of the working set size.

To do:
- better names for the new fields
- proper locking when gathering stats
- some code refactoring for the new stats and triiming code
- use boolean_t where appropriate