sdt/amd64: implement probe calls similarly to fbt
Previously on all platforms there would be a conditional function call
for each probe.  That means that every time we would check if a particular
probe is enabled.  Now on amd64 only the actual call instructions are
replaced with NOPs, so the overhead of a disabled probe is arranging
probe arguments into the proper registers (and onto the stack if there are
more than 6 of them) and executing 5 NOP instructions.  When the probe is
enabled one of the NOPs is replaced with the LOCK prefix opcode resulting
in a LOCK NOP instruction, which is invalid and generates #UD fault.
The fault is handled similarly to #BP trap, but there is a difference in
the value of %rip when exceptions are generated.
Another difference with FBT is that #BP is generated after a CALL
instruction when entering a callee, thus a return address is pushed onto
the stack potentially after the 7+th call arguments.  With SDT #UD is
generated instead of the CALL, so there is no return address there.

The call sites of the probes have to be patched with NOPs before
the code is executed, this is done in kern_sdt during SI_SUB_KLD
initialization phase.  The sdt module handles modules added after it's
loaded.

The call sites of the probes are determined by using inline assembly.
This is in contrast with illumos / Solaris where a special support from
linker is used and the call sites are recorded as relocations.
The FreeBSD method is much less reliable because a compiler is free to
rearrange the inline assembly blocks and the C function calls.
Unfortunately, adding special link time support for the probe call sites
is quite cumbersome and keeping all relocation information (ld -q)
significantly increases size of the kernel and modules.

To do:
- fix a window of unpatched SDT probes in modules loaded after boot but
  before sdt module is loaded
- replace magic values (e.g. instruction opcodes) with proper constants
- consider using #BP instead of #UD on amd64, there should be not reason
  not to
- cleanly separate platform specific code
- generalize dtrace_getarg() so that it can used by both FBT and SDT, a new
  parameter for an offset of the no-stack argument from saved %rsp is
  needed
- drive SDT probe creation by call sites rather than by probe definitions,
  this would allow to automatically populate function name attribute in
  the case where probe definition does not specify a predefined function
  name; if the same statically defined probe is used in several functions
  then a different probe would be created for each function with an
  appropriate function name attribute
- replace inline assembly with a build and toolchain changes; for example,
  we could initially keep all relocation information and then strip it, so
  that only relocations for SDT calls are kept (note that we already
  keep all relocations for amd64 kernel loadable modules and the
  relocations are applied at loading time, this is in contrast to the
  kernel and to the modules on other platforms)